// src/app/models/company-info.ts
export class CompanyInfo {
        id:             number;
        companyName:    string;
        rtn:            string;
        addressLine1:   string;
        addressLine2:   string;
        phone1:         string;
        phone2:         string;
        email2:         string;
        email1:         string;
        userId:         number;
        userName: string;
        createDate:     Date;
        fileId: number;
        taxValue: number;
        negativeInventory: boolean;
        printLetter: boolean;
        path: string;
        extension   : string;
}